agent-core-py/requirements.txt
groq>=0.11.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langgraph>=0.2.0
langchain-groq>=0.2.0
langchain-core>=0.2.27
motor>=3.6.0
neo4j>=5.25.0
qdrant-client>=1.12.0
pydantic>=2.9.0
python-dotenv>=1.0.1
httpx>=0.27.0
